Add spec coverage for possible exposure indicator
Why:
Currently the feature of showing a 'Next Steps' button on the
ExposureHistory screen has not test coverage. As this is an important
part of the user flow, we would like to add some basic test coverage.
