+++ app/views/ExposureHistory/ExposureDatumDetail.tsx
@@ -63,6 +63,7 @@ const PossibleExposureDetail = ({
       </View>
       <View style={styles.ctaContainer}>
         <TouchableOpacity
+          testID={'exposure-history-next-steps-button'}
           style={styles.nextStepsButton}
           onPress={handleOnPressNextSteps}>
           <Typography style={styles.nextStepsButtonText}>
